requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
cycler==0.10.0
Django==3.2.8
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-model-utils==4.2.0
django-phonenumber-field==5.2.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
future==0.18.2
html5lib==1.1
kiwisolver==1.3.2
matplotlib==3.4.3
mysqlclient==2.0.3
numpy==1.21.2
phonenumbers==8.12.35
Pillow==8.4.0
pycparser==2.20
pymongo==3.12.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.2
six==1.16.0
sqlparse==0.2.4
webencodings==0.5.1
xhtml2pdf==0.2.5
